Removing MacOS file.

+++ include/nola/cxx17/root.hxx
@@ -6,7 +6,7 @@
 
 #include <cassert> // assert
 #include <cmath>   // std::abs
-#include <cstdlib> // std::size_t
+#include <cstdint> // std::int32_t
 #include <limits>  // std::numeric_limits
 #include <utility> // std::forward
 #include <nola/root_impl.hxx>
@@ -23,47 +23,47 @@ namespace root
 
 
 template <class UnaryFunction, class Real>
-inline std::tuple<Real, std::size_t>
+inline std::tuple<Real, std::int32_t>
 bisection(UnaryFunction && f, Real a, Real b);
 
 template <class UnaryFunction, class Real>
-inline std::tuple<Real, std::size_t>
+inline std::tuple<Real, std::int32_t>
 bisection(UnaryFunction && f, Real a, Real b, Real tol, std::int32_t maxiter);
 
 
 template <class UnaryFunction, class Real>
-inline std::tuple<Real, std::size_t>
+inline std::tuple<Real, std::int32_t>
 brent(UnaryFunction&& f, Real x0, Real x1);
 
 template <class UnaryFunction, class Real>
-inline std::tuple<Real, std::size_t>
+inline std::tuple<Real, std::int32_t>
 brent(UnaryFunction&& f, Real x0, Real x1, Real tol, std::int32_t maxiter);
 
 
 template <class UnaryFunction, class Real>
-inline std::tuple<Real, std::size_t>
+inline std::tuple<Real, std::int32_t>
 fixed_point(UnaryFunction&& g, Real x0);
 
 template <class UnaryFunction, class Real>
-inline std::tuple<Real, std::size_t>
+inline std::tuple<Real, std::int32_t>
 fixed_point(UnaryFunction&& g, Real x0, Real tol, std::int32_t maxiter);
 
 
 template <class UnaryFunction1, class UnaryFunction2, class Real>
-inline std::tuple<Real, std::size_t>
+inline std::tuple<Real, std::int32_t>
 newton(UnaryFunction1&& f, UnaryFunction2&& fp, Real x0);
 
 template <class UnaryFunction1, class UnaryFunction2, class Real>
-inline std::tuple<Real, std::size_t>
+inline std::tuple<Real, std::int32_t>
 newton(UnaryFunction1&& f, UnaryFunction2&& fp, Real x0, Real tol, std::int32_t maxiter);
 
 
 template <class UnaryFunction, class Real>
-inline std::tuple<Real, std::size_t>
+inline std::tuple<Real, std::int32_t>
 secant(UnaryFunction&& f, Real x0, Real x1);
 
 template <class UnaryFunction, class Real>
-inline std::tuple<Real, std::size_t>
+inline std::tuple<Real, std::int32_t>
 secant(UnaryFunction&& f, Real x0, Real x1, Real tol, std::int32_t maxiter);
 
 
@@ -75,14 +75,14 @@ secant(UnaryFunction&& f, Real x0, Real x1, Real tol, std::int32_t maxiter);
 // Bisection Method
 
 template <class Unary_function, class Real>
-inline std::tuple<Real, std::size_t>
+inline std::tuple<Real, std::int32_t>
 bisection(Unary_function && f, Real a, Real b, Real tol, std::int32_t maxiter)
 {
   return detail::bisection_impl( std::forward<Unary_function>(f), a, b, tol, maxiter );
 }
 
 template <class Unary_function>
-inline std::tuple<float, std::size_t>
+inline std::tuple<float, std::int32_t>
 bisection(Unary_function && f, float a, float b)
 {
   float        tol{1e-5};
@@ -91,7 +91,7 @@ bisection(Unary_function && f, float a, float b)
 }
 
 template <class Unary_function>
-inline std::tuple<double, std::size_t>
+inline std::tuple<double, std::int32_t>
 bisection(Unary_function && f, double a, double b)
 {
   double       tol{1e-10};
